+++ libftprintf/ft_strlen.c
@@ -4,9 +4,9 @@
 ** ft_strlen: возвращает количество символов в строке.
 */
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t i;
+	int i;
 
 	i = 0;
 	while (s[i] != '\0')
